# railway-backend/openapi.yaml
openapi: 3.0.3
info:
  title: All For Pets — Municipal Pet Registry API
  version: 1.0.0
  description: |
    REST API powering the AFP web + Expo mobile app.

    ### Authentication
    - `POST /api/auth/login` returns `token` (JWT access, 15 min) and
      `refreshToken` (opaque, device-bound, 1–30 days).
    - Send `Authorization: Bearer <token>` on every protected call.
    - On 401 → call `POST /api/auth/refresh` with the refresh token; a **new**
      refresh token is returned (rotation) and the old one is revoked.
    - Mobile clients should send `X-Device-Label` and `X-Expo-Push-Token`
      headers on login/register/refresh so we can list / revoke sessions and
      push notifications.

    ### Caching
    - Public `/api/geo/*` endpoints return `ETag` + `Cache-Control`.
      Send `If-None-Match` to get a 304 with an empty body.

    ### File uploads (mobile)
    - Prefer the `/upload-photo/sign` + `/upload-photo/confirm` flow so the
      binary goes directly to S3/Azure/Cloudinary without touching the API.
servers:
  - url: https://afp.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local development (Node)
...
security:
  - bearerAuth: []
paths:
...
  /api/pets:
    post:
      tags: [Pets]
      summary: Register a new pet (pending admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, species]
              properties:
                name:        { type: string }
                species:     { type: string, enum: [dog, cat, rabbit, bird, other] }
                breed:       { type: string, nullable: true }
                colour:      { type: string, nullable: true }
                gender:      { type: string, enum: [male, female], nullable: true }
                dateOfBirth: { type: string, format: date, nullable: true }
                paymentId:   { type: string, nullable: true }
                txnRef:      { type: string, nullable: true }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Pet" } } }
...
  /api/doctors:
    get:
      tags: [Doctors]
      security: []
      parameters:
        - { in: query, name: cityId, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Doctor" } } } }
  /api/shops:
    get:
      tags: [Shops]
      security: []
      parameters:
        - { in: query, name: cityId, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Shop" } } } }

  # ───────────── Reports ─────────────
  /api/reports:
    post:
      tags: [Reports]
      summary: File a report (stray, lost, unlicensed, cruelty)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, address, mobile]
              properties:
                type:    { type: string, enum: [stray, lost, unlicensed, cruelty] }
                address: { type: string }
                mobile:  { type: string }
                notes:   { type: string, nullable: true }
      responses:
        "201":
          description: Created

  # ───────────── Discussions ─────────────
  /api/discussions:
    get:
      tags: [Discussions]
      responses:
        "200": { description: OK }
    post:
      tags: [Discussions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:    { type: string }
                body:     { type: string }
                category: { type: string, default: general }
      responses:
        "201": { description: Created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Pet:
      type: object
      properties:
        id:                  { type: integer }
        pet_id:              { type: string, description: "Human-readable ID, e.g. AFP-JP-0004" }
        owner_id:            { type: integer }
        name:                { type: string }
        species:             { type: string }
        breed:               { type: string, nullable: true }
        colour:              { type: string, nullable: true }
        gender:              { type: string, nullable: true }
        date_of_birth:       { type: string, format: date, nullable: true }
        photo_url:           { type: string, nullable: true }
        certificate_url:     { type: string, nullable: true }
        registration_status: { type: string, enum: [pending, approved, rejected] }
        licence_status:      { type: string, nullable: true }
        licence_expiry_date: { type: string, format: date, nullable: true }
        breeding_opt_in:     { type: boolean }
        owner_name:          { type: string, nullable: true }
        city_name:           { type: string, nullable: true }
        nigam_name:          { type: string, nullable: true }
        ward_number:         { type: string, nullable: true }
        created_at:          { type: string, format: date-time }
...
    Doctor:
      type: object
      properties:
        id:             { type: integer }
        name:           { type: string }
        qualification:  { type: string, nullable: true }
        specialization: { type: string, nullable: true }
        clinic_name:    { type: string, nullable: true }
        address:        { type: string, nullable: true }
        mobile:         { type: string, nullable: true }
        timings:        { type: string, nullable: true }
        is_24hr:        { type: boolean }
        city_name:      { type: string, nullable: true }
    Shop:
      type: object
      properties:
        id:         { type: integer }
        name:       { type: string }
        owner_name: { type: string, nullable: true }
        speciality: { type: string, nullable: true }
        address:    { type: string, nullable: true }
        mobile:     { type: string, nullable: true }
        timings:    { type: string, nullable: true }
        city_name:  { type: string, nullable: true }
...